Created Item objects from form data.
I used the form data to create an Item object and added it to an order array. The HTML form had to be changed a little bit in order to get it to work. 

Unfortunately, I also introduced a bug that adds the extras to the items in both the menu and the order arrays which could cause problems further down the line.